mptcp: Take meta-lock when creating secondary subflow
Unfortunately, there are a number of problems of making the whole process lockless.

When we do not take the lock, it means that we could be holding a
subflow lock without holding the meta-lock during the creation-process.
That means, if we enter mptcp_disconnect() on another CPU (e.g., from
inet_csk_listen_stop()) we would drop the subflow's lock. That means, we
effectively have a CPU working on the subflow without holding any locks.
The reason for some refcount-warnings that pop up.

Also, we had lock-ordering problems. When the master was created we first
took the meta-lock and then the master lock. When a child was created we
first were holding the child's lock and then the meta-lock in
tcp_child_process(). LOCKDEP does not like this.

This patch makes sure that we *always* hold the meta-lock when creating
a subflow (whether master or not). For the master-case this is already
guaranteed through the normal flow of events. For the child-case it is
through the new code in tcp_check_req().

Fixes: 4966859c2ad3 ("mptcp: Don't take meta-lock when receiving third ACK")
Signed-off-by: Christoph Paasch <cpaasch@apple.com>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
(cherry picked from commit b4563ba68a6f3b686cc14ccc3ce85978fc439cad)
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>