mptcp: Make sure that we don't overfill subflows
The tests in mptcp_is_temp_unavailable() will just tell the scheduler
whether a single packet fits in the send-queue. Later on in
mptcp_next_segment we then size the packet so that it fits. We were only
checking tcp_cwnd_test, which does not take into account the amount of
data that was already queued but not yet sent (due to TSQ).

We need to take this correctly into account by looking at write_seq.

Further, our logic with "needed" was wrong. We really want to take the
minimum to make sure that the segment we schedule gets sent right away.

Reported-by: Anh Vu Vu <vuanh.vu@ikt.uni-hannover.de>
Fixes: Zero-day bug
Signed-off-by: Christoph Paasch <cpaasch@apple.com>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
(cherry picked from commit 988ec1373becf376cb44e66fd1c60af4282a7f84)
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>